% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper-package.R
\docType{package}
\name{bootstrapper-package}
\alias{bootstrapper-package}
\title{bootstrapper: Opinionated R Package Quickstart}
\description{
Quickly start new R packages, adding some opinionated default GHA, suggested extensions, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://visruthsk.github.io/bootstrapper/}
}

}
\author{
\strong{Maintainer}: Visruth Srimath Kandali \email{public@visruth.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID}) [copyright holder]

}
\keyword{internal}
