% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Bootstrap a New R Package}
\usage{
bootstrapper(
  fields = getOption("usethis.description"),
  setup_gha = TRUE,
  setup_air = TRUE,
  setup_jarl = TRUE,
  setup_dependabot = TRUE,
  setup_AGENTS = FALSE,
  setup_precommit = TRUE,
  setup_touchstone = FALSE,
  setup_touchstone_plots = FALSE,
  ...
)
}
\arguments{
\item{fields}{Named list of \code{DESCRIPTION} fields passed to
\code{\link[usethis:create_package]{usethis::create_package()}}. See \code{\link[usethis:use_description]{usethis::use_description()}}}

\item{setup_gha}{Whether to configure GitHub Actions setup.}

\item{setup_air}{Whether to configure Air formatting.}

\item{setup_jarl}{Whether to configure Jarl linting.}

\item{setup_dependabot}{Whether to write a Dependabot configuration.}

\item{setup_AGENTS}{Whether to write a default AGENTS file.}

\item{setup_precommit}{Whether to write a Bash pre-commit hook.}

\item{setup_touchstone}{Whether to setup Touchstone benchmarking.}

\item{setup_touchstone_plots}{Whether to use the Touchstone comment workflow
that publishes benchmark plots to a separate branch. Only used when
\code{setup_touchstone = TRUE}.}

\item{...}{Additional arguments passed to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Create a package with some opinionated setup.
}
