% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twogroup.R
\name{plot.twogroup}
\alias{plot.twogroup}
\title{Plot}
\usage{
\method{plot}{twogroup}(tg, use = NULL, alpha = 0.05)
}
\arguments{
\item{tg}{A `twogroup` object}
}
\value{
none
}
\description{
Plots the resampling distribution for different hypotheses
}

