% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{setup_dependabot}
\alias{setup_dependabot}
\title{Configure Dependabot Defaults}
\usage{
setup_dependabot()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Writes a default Dependabot configuration for GitHub Actions.
}
