% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{setup_gha}
\alias{setup_gha}
\title{Configure GitHub Actions Defaults}
\usage{
setup_gha()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Sets up standard GitHub Actions used by this package template and updates
workflow references.
}
