% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{setup_precommit}
\alias{setup_precommit}
\title{Configure Pre-Commit Hook}
\usage{
setup_precommit()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Writes a Bash pre-commit hook that runs format and lint checks.
}
