% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{setup_touchstone}
\alias{setup_touchstone}
\title{Configure Touchstone}
\usage{
setup_touchstone(plots = FALSE)
}
\arguments{
\item{plots}{Whether to use the workflow which writes touchstone plots (and needs more permissions).}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Write a modified Touchstone GHA to benchmark PRs. You still
need to write an appropriate \code{script.R} for the actual
benchmarks. This version of the touchstone commenting GHA
updates a single comment instead of making multiple. Optionally
also adds the touchstone plots in a dropdown--these plots are
stored on a new branch.
}
