% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twogroup.R
\name{stattable}
\alias{stattable}
\title{Tables of Statistics}
\usage{
stattable(tg)
}
\arguments{
\item{tg}{A `twogroup` object}
}
\value{
A list of 3 tables
}
\description{
Creates a table of p-values across methods, as well as confidence intervals, and degrees of freedom
}

