% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twogroup.R
\name{summary.twogroup}
\alias{summary.twogroup}
\title{Summary}
\usage{
\method{summary}{twogroup}(tg)
}
\arguments{
\item{tg}{A `twogroup` object}
}
\value{
none
}
\description{
Prints a summary of statistical significance and confidence intervals
}

