% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twogroup.R
\name{twogroup}
\alias{twogroup}
\title{Two Group Comparison}
\usage{
twogroup(X, Y, nboot = 10000, replace = T, location = mean.fast,
  scale = sd.fast, dither = F, amp = 0.7, use = c("equal.var",
  "unequal.var", "rank", "Student", "Welch", "Wilcox", "ES"))
}
\arguments{
\item{X}{The first group}

\item{Y}{The second group}

\item{nboot}{The number of resamples}

\item{replace}{Whether to sample with replacement}

\item{location}{Function that measures central location. Defaults to mean}

\item{scale}{Function that measures spread. Defaults to standard deviation}

\item{dither}{Whether to add dithering noise. Defaults to no}

\item{amp}{Amplitude of dithering noise.}

\item{use}{Statistics to compute. Can use multiple. Defaults to all.}
}
\value{
A `twogroup` object
}
\description{
Computes statistical significance using parametric, non-parametric, and bootstrap methods
}

