% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{use_license}
\alias{use_license}
\title{Choose and Apply a License}
\usage{
use_license()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Prompts for a license choice in interactive sessions and applies the selected
\code{usethis} license helper.
}
