% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper-package.R
\docType{package}
\name{bootstrapper-package}
\alias{bootstrapper-package}
\title{bootstrapper: Opinionated R Package Quickstart}
\description{
Bootstrap new R packages with strongly opinionated defaults. Creates the package scaffold, wires up tests, README and NEWS, configures GitHub Actions and Dependabot, adds Air/Jarl setup, and drop in a default AGENTS.md for building R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://visruthsk.github.io/bootstrapper/}
}

}
\author{
\strong{Maintainer}: Visruth Srimath Kandali \email{public@visruth.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID}) [copyright holder]

}
\keyword{internal}
