% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Bootstrap a New R Package}
\usage{
bootstrapper(path = ".", fields, private = TRUE, ...)
}
\arguments{
\item{path}{Path where the package should be created. Defaults to \code{"."}}

\item{fields}{Named list of \code{DESCRIPTION} fields passed to
\code{\link[usethis:create_package]{usethis::create_package()}}. See \code{\link[usethis:use_description]{usethis::use_description()}}}

\item{private}{Whether to create the GitHub repository as private. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Create a package with some opinionated setup.
}
