% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Bootstrap a New R Package}
\usage{
bootstrapper(
  path = ".",
  fields = getOption("usethis.description", list(`Authors@R` = person("Visruth",
    "Srimath Kandali", , "public@visruth.com", role = c("aut", "cre", "cph"), comment =
    c(ORCID = "0009-0005-9097-0688")))),
  private = TRUE,
  setup_gha = TRUE,
  setup_dependabot = TRUE,
  setup_AGENTS = FALSE,
  setup_precommit = TRUE,
  ...
)
}
\arguments{
\item{path}{Path where the package should be created. Defaults to \code{"."}}

\item{fields}{Named list of \code{DESCRIPTION} fields passed to
\code{\link[usethis:create_package]{usethis::create_package()}}. See \code{\link[usethis:use_description]{usethis::use_description()}}}

\item{private}{Whether to create the GitHub repository as private. Defaults to \code{TRUE}.}

\item{setup_gha}{Whether to configure GitHub Actions setup.}

\item{setup_dependabot}{Whether to write a Dependabot configuration.}

\item{setup_AGENTS}{Whether to write a default AGENTS file.}

\item{setup_precommit}{Whether to write a Bash pre-commit hook.}

\item{...}{Additional arguments passed to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Create a package with some opinionated setup.
}
