% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{create_package}
\alias{create_package}
\title{Create a Package and Connect GitHub}
\usage{
create_package(fields = getOption("usethis.description"), ...)
}
\arguments{
\item{fields}{Named list of \code{DESCRIPTION} fields passed to
\code{\link[usethis:create_package]{usethis::create_package()}}. See \code{\link[usethis:use_description]{usethis::use_description()}}}

\item{...}{Additional arguments passed to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Create a package in root, prompts for a license, cleans
up build ignore file.
}
