% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{setup_formatter}
\alias{setup_formatter}
\title{Configure Formatter Defaults}
\usage{
setup_formatter(air = TRUE, jarl = TRUE, format = TRUE, gha = TRUE)
}
\arguments{
\item{air}{Whether to configure Air formatting.}

\item{jarl}{Whether to configure Jarl linting.}

\item{format}{Whether to format the repository after configuring tools.}

\item{gha}{Whether to create the appropriate GHA for automatic formatting on PRs.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Sets up Air and/or Jarl formatting defaults. Optionally
format the repository once.
}
