% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{setup_touchstone}
\alias{setup_touchstone}
\title{Configure Touchstone}
\usage{
setup_touchstone()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Write a modified Touchstone GHA to benchmark PRs. You still
need to write an appropriate \code{script.R} for the actual
benchmarks. This version of the touchstone commenting GHA
updates a single comment instead of making multiple, and also
adds the touchstone plots in a dropdown. These are stored on a new
branch.
}
