% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twogroup.R
\name{twogroup}
\alias{twogroup}
\title{Two Group Comparison}
\usage{
twogroup(X, Y, nboot = 10000, replace = T, location = mean.fast,
  scale = sd.fast, dither = F, amp = 0.7, use = c("equal.var",
  "unequal.var", "rank", "Student", "Welch", "Wilcox", "ES"))
}
\arguments{
\item{X}{The first group}

\item{Y}{The second group}

\item{nboot}{The number of resamples}

\item{replace}{Whether to sample with replacement}

\item{location}{Function that measures central location. Defaults to mean}

\item{scale}{Function that measures spread. Defaults to standard deviation}

\item{dither}{Whether to add dithering noise. Defaults to no}

\item{amp}{Amplitude of dithering noise.}

\item{use}{Statistics to compute. Can use multiple. Defaults to all.}
}
\value{
A `twogroup` object
}
\description{
Computes statistical significance using parametric, non-parametric, and bootstrap methods
}
\examples{

# create dataset
n <- 30
X <- rnorm(n,0,1)
Y <- rnorm(n,1,1)

# run
tg <- twogroup(X,Y)
summary(tg)
plot(tg)
plot(tg, use="equal.var")
plot(tg, use="unequal.var")
plot(tg, use="rank")
plot(tg, use="ES")

# Compute false positive rates for some experimental design
n1 <- 5; n2 <- 15
nsim <- 300
X <- matrix(rnorm(n1*nsim, mean=0, sd=3), nrow=nsim)
Y <- matrix(rnorm(n2*nsim, mean=0, sd=1), nrow=nsim)
simdat <- cbind(X,Y)

sims <- apply(simdat, 1, FUN=function(x) twogroup(x[1:n1], x[(n1+1):(n1+n2)], nboot=300))
ptabs <- lapply(sims, function(x) stattable(x)$p)
fpr <- rowSums(sapply(ptabs, function(x) x<.05))/nsim
names(fpr) <- rownames(ptabs[[1]])
cbind(fpr)

}

